Added base classes to all mixins.
Mixins are being checked for functionality in many places.  Originally
I made them just to make it easy to reuse the same slots, but now its
become more of a "add this functionality" thing.  As a result, using a
base class for the functionality exposure is more useful, as it "does
this class have this functionality" is now more generic, rather than
tied to the specific implementation.  I can easily see needing to make
other types of some of these functionalities, but not using a base to
test for them prevented alternative implementations.  This "fixes"
that weakness, I believe.